Adding new errors and improving the swagger

+++ src/main/java/com/example/emailsender/services/TokenService.java
@@ -1,10 +1,16 @@
 package com.example.emailsender.services;
 
-import com.example.emailsender.entities.EmailReceiver;
+import com.example.emailsender.controllers.handler.ExpiredTokenException;
+import com.example.emailsender.controllers.handler.ExistingTokenException;
+import com.example.emailsender.controllers.handler.TokenNotFoundException;
+import com.example.emailsender.entities.Token;
 import com.example.emailsender.repositories.TokenRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
+import java.time.LocalTime;
+import java.time.ZoneId;
 import java.util.Objects;
 import java.util.Random;
 
@@ -17,11 +23,18 @@ public class TokenService {
     public String insertToken(final String userEmail){
         final var existentToken = repository.findByEmail(userEmail);
         if(Objects.nonNull(existentToken)){
-            repository.delete(existentToken);
+            if(Instant.now().getEpochSecond() - existentToken.getDate().getEpochSecond() > 3600) {
+                repository.delete(existentToken);
+            } else {
+                LocalTime d = existentToken.getDate().plusSeconds(3600)
+                        .atZone(ZoneId.of("America/Sao_Paulo")).toLocalTime();
+                throw new ExistingTokenException("Seu token expira as: " + d.toString(), existentToken.getEmail());
+            }
         }
-        final var token = EmailReceiver.builder()
+        final var token = Token.builder()
                 .email(userEmail)
                 .token(generate())
+                .date(Instant.now())
                 .build();
 
         repository.insert(token);
@@ -29,16 +42,20 @@ public class TokenService {
         return token.getToken();
     }
 
-    public boolean validateToken(final EmailReceiver emailReceiverSrc) {
+    public void validateToken(final Token tokenSrc) {
 
-        final var tokenFnd = repository.findByEmailAndToken(emailReceiverSrc.getEmail(), emailReceiverSrc.getToken());
+        final var tokenFnd = repository.findByEmailAndToken(tokenSrc.getEmail(), tokenSrc.getToken());
 
         if(Objects.nonNull(tokenFnd)){
-            repository.delete(tokenFnd);
-            return true;
+            if(Instant.now().getEpochSecond() - tokenFnd.getDate().getEpochSecond() < 3600) {
+                repository.delete(tokenFnd);
+            } else {
+                throw new ExpiredTokenException("Por favor, gere um novo token");
+            }
+        } else {
+            throw new TokenNotFoundException("Não foi encontrado o token " + tokenSrc.getToken()
+                    + " associado com o email " + tokenSrc.getEmail());
         }
-
-        return false;
     }
